Login dialog fix.
When the dialog box is presented to the user (in the fringe case where the credentials stored away in the keychain have changed), we don't want the user to log in with a different account. So the dialog has been changed to only work if the user enters the same username that was used to set up the device in the first place.